refactor: move joinSections to src/util/, fix filtered section blank line
- Relocates joinSections to src/util/markdown.mts to avoid cross-layer
  import from reporters/ into cli/
- Updates test to match new per-section trim semantics (internal newlines
  preserved unchanged)
- Restores blank line after ### Filtered heading in reporters/text.mts
